backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-generativeai==0.8.3
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.5
numpy==1.26.4
scikit-learn==1.5.2
pydantic==2.9.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pgvector==0.3.5
firebase-admin==6.5.0
